Split NodeDetailsSection
